requirements.txt
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
filelock==3.13.3
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
haversine==2.8.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
requests-file==2.0.0
setuptools==69.2.0
soupsieve==2.5
tldextract==5.1.2
urllib3==2.2.1
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.2
